+++ dev_dependencies.txt
@@ -1,0 +1,3 @@
+pytest
+pytest-asyncio
+pytest-mock
