+++ tests/authentication_msal/test_msal_auth.py
