+++ tests/authentication_msal/test_msal_connection_manager.py
