+++ tests/hosting_core/test_activity_handler.py
