+++ tests/hosting_core/test_auth_configuration.py
