+++ tests/hosting_core/test_error_handling.py
