+++ tests/hosting_core/test_state.py
