+++ tests/hosting_core/tools/testing_adapter.py
